Enforce limits on the classic and block editor
